[Refactor] #150 PlanRepository에 통합
