+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/datasource/PlanDataSource.kt
@@ -1,6 +1,7 @@
 package kr.tekit.lion.daongil.data.datasource
 
 import kr.tekit.lion.daongil.data.dto.remote.response.plan.OpenPlanListResponse
+import kr.tekit.lion.daongil.data.dto.remote.response.plan.PlaceSearchResultsResponse
 import kr.tekit.lion.daongil.data.network.service.PlanService
 import okhttp3.RequestBody
 
@@ -14,6 +15,9 @@ class PlanDataSource(
 
     suspend fun addNewPlan(request: RequestBody){
         planService.addNewPlan(request)
+    }
+    suspend fun getPlaceSearchResult(word: String, page: Int, size: Int) : PlaceSearchResultsResponse {
+        return planService.getPlaceSearchResults(word, page, size)
     }
 
 }
