+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/repository/PlanRepositoryImpl.kt
@@ -4,6 +4,7 @@ import kr.tekit.lion.daongil.data.datasource.PlanDataSource
 import kr.tekit.lion.daongil.data.dto.remote.request.toRequestBody
 import kr.tekit.lion.daongil.domain.model.NewPlan
 import kr.tekit.lion.daongil.domain.model.OpenPlan
+import kr.tekit.lion.daongil.domain.model.PlaceSearchResult
 import kr.tekit.lion.daongil.domain.repository.PlanRepository
 
 class PlanRepositoryImpl(
@@ -15,5 +16,14 @@ class PlanRepositoryImpl(
 
     override suspend fun addNewPlan(request: NewPlan) {
         return planDataSource.addNewPlan(request.toRequestBody())
+    }
+
+    // Repository Interface에 구현한 getPlaceSearchResult 함수
+    override suspend fun getPlaceSearchResult(
+        word: String,
+        page: Int,
+        size: Int
+    ): PlaceSearchResult {
+        return planDataSource.getPlaceSearchResult(word, page, size).toDomainModel()
     }
 }
