+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/presentation/scheduleform/vm/ScheduleFormViewModelFactory.kt
@@ -3,7 +3,6 @@ package kr.tekit.lion.daongil.presentation.scheduleform.vm
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import kr.tekit.lion.daongil.domain.repository.PlaceDetailInfoRepository
-import kr.tekit.lion.daongil.domain.repository.PlaceSearchResultRepository
 import kr.tekit.lion.daongil.domain.repository.PlanRepository
 import kr.tekit.lion.daongil.domain.usecase.AddNewPlanUseCase
 import kr.tekit.lion.daongil.domain.usecase.GetPlaceDetailInfoUseCase
@@ -13,7 +12,7 @@ import kr.tekit.lion.daongil.domain.usecase.GetPlaceSearchResultUseCase
 // ViewModelProvider.Factory 인터페이스를 구현한다.
 class ScheduleFormViewModelFactory : ViewModelProvider.Factory {
     private val getPlaceSearchResultUseCase = GetPlaceSearchResultUseCase(
-        PlaceSearchResultRepository.create() // repository의 인스턴스를 생성하여 전달
+        PlanRepository.create() // repository의 인스턴스를 생성하여 전달
     )
     private val getPlaceDetailInfoUseCase = GetPlaceDetailInfoUseCase(
         PlaceDetailInfoRepository.crate()
